requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
discord.py==1.6.0
idna==2.10
joblib==1.0.0
mccabe==0.6.1
multidict==5.1.0
praw==7.1.4
prawcore==1.5.0
nltk==3.5
python-dotenv==0.15.0
regex==2020.11.13
requests==2.25.1
tqdm==4.56.0
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.26.3
websocket-client==0.57.0
yarl==1.6.3